Improve cluster leave logic and rename some cluster internal interfaces
