+++ cluster/src/test/java/io/scalecube/cluster/membership/DummyMembershipProtocol.java
@@ -12,7 +12,7 @@ import rx.Observable;
 /**
  * @author Anton Kharenko
  */
-public class DummyMembershipProtocol implements IMembershipProtocol {
+public class DummyMembershipProtocol implements MembershipProtocol {
 
   private Member localMember;
   private List<Member> remoteMembers = new ArrayList<>();
